move tools/stellar-archivist/historyarchive -> support/historyarchive
