+++ support/historyarchive/hash.go
