+++ support/historyarchive/hash_test.go
