+++ support/historyarchive/scan.go
