+++ support/historyarchive/verify.go
